Delay MenuHandler by 2 ticks
To play well with other mods manipulating the `MenuPage._materialContainers` list, i.e., giving other mods enough time to add to / remove from that list, we delay the instantiation of the MenuHandler by 2 ticks = 1/30 seconds.

I think a 2-tick delay is enough. Not too long that it annoys players, but not too short that other mods don't have the chance to finish their manipulations.