Optimize iteration in OnStackAmountEntered()
First, we implement the original TODO: Loop only as many as necessary. This is done by grabbing the maximum possible amount in the OpenSplitMenu() method

Then we play sound only during the first iteration; not only should this make the loop faster, but also removes the horrible LOUD BRAP noise due to a sound clip being layered several times on top of itself with millisecond offset inbetween.

Unfortunately, a bug in the game code forces us to do things in a roundabout way: Nullify Game1.soundBank for the second loop onwards, restoring it on every loop to prevent sound not playing (if needed by other procedures)