Remove "cooking" variable because it's no longer used
`_materialContainers` will be populated by the game itself with the proper additional containers (i.e., fridges for cooking, attached chests for Workbench).

Even mods that 'extends' the reach of the Stove/Workbench do it by adding additional containers to `_materialContainers`

Because of this, detecting whether the page is 'cooking' or not is no longer necessary.